Rearranged DMapUtilties into their eventual module location, added CPP for mkDynPure test.
