+++ test/Reflex/Test/TH.hs
@@ -13,7 +13,7 @@ testCases =
     testE "mkDynPure" $ do
       dyn1 <- holdDyn 0 =<< events1
       dyn2 <- holdDyn 0 =<< events2
-      let dynResult = [mkDynPure|$dyn1 + $dyn2|]
+      let dynResult = [mkDynPure|($dyn1, $dyn2)|]
       return $ updated dynResult
   ] where
